Move infraestructure code to lib
