+++ lib/turbo-rails.rb
@@ -1,8 +1,4 @@
-require "turbo/engine"
-
 module Turbo
-  extend ActiveSupport::Autoload
-
   class << self
     attr_writer :signed_stream_verifier_key
 
@@ -15,3 +11,5 @@ module Turbo
     end
   end
 end
+
+require_relative "load_turbo"
