+++ lib/turbo/helpers/helpers.rb
@@ -1,0 +1,7 @@
+module Turbo::Helpers
+  include Turbo::DriveHelper
+  include Turbo::FramesHelper
+  include Turbo::IncludesHelper
+  include Turbo::Streams::ActionHelper
+  include Turbo::StreamsHelper
+end
