+++ turbo-rails.gemspec
@@ -11,6 +11,7 @@ Gem::Specification.new do |s|
 
   s.required_ruby_version = ">= 2.6.0"
   s.add_dependency "rails", ">= 6.0.0"
+  s.add_dependency "zeitwerk", "~> 2.4"
 
   s.files      = `git ls-files`.split("\n")
   s.test_files = `git ls-files -- test/*`.split("\n")
